feat: Moved a logic that calculate a pose to the system.
